Making a new TreeGrid component. Separating treegrid logic from tableDataView
