+++ src/sql/base/browser/ui/table/tableDataView.ts
@@ -9,7 +9,6 @@ import { compare as stringCompare } from 'vs/base/common/strings';
 
 import { FilterableColumn } from 'sql/base/browser/ui/table/interfaces';
 import { IDisposableDataProvider } from 'sql/base/common/dataProvider';
-import { treeGridExpandableColumnFormatter } from 'sql/base/browser/ui/table/formatters';
 
 export interface IFindPosition {
 	col: number;
@@ -21,7 +20,7 @@ export type TableFilterFunc<T extends Slick.SlickData> = (data: Array<T>, column
 export type TableSortFunc<T extends Slick.SlickData> = (args: Slick.OnSortEventArgs<T>, data: Array<T>) => Array<T>;
 export type TableFindFunc<T extends Slick.SlickData> = (val: T, exp: string) => Array<number>;
 
-function defaultCellValueGetter(data: any): any {
+export function defaultCellValueGetter(data: any): any {
 	return data;
 }
 
@@ -55,7 +54,7 @@ export function defaultSort<T extends Slick.SlickData>(args: Slick.OnSortEventAr
 	return data.sort((a, b) => comparer(a, b) * sign);
 }
 
-function defaultFilter<T extends Slick.SlickData>(data: T[], columns: FilterableColumn<T>[], cellValueGetter: CellValueGetter = defaultCellValueGetter): T[] {
+export function defaultFilter<T extends Slick.SlickData>(data: T[], columns: FilterableColumn<T>[], cellValueGetter: CellValueGetter = defaultCellValueGetter): T[] {
 	let filteredData = data;
 	columns?.forEach(column => {
 		if (column.filterValues?.length > 0 && column.field) {
@@ -64,20 +63,6 @@ function defaultFilter<T extends Slick.SlickData>(data: T[], columns: Filterable
 			});
 		}
 	});
-
-	const isTreeGrid = columns.find(c => c.formatter === treeGridExpandableColumnFormatter);
-	if (isTreeGrid) {
-		filteredData = filteredData.filter((item) => {
-			let parent = data[item.parent];
-			while (parent) {
-				if (!parent.expanded) {
-					return false;
-				}
-				parent = data[parent.parent];
-			}
-			return true;
-		});
-	}
 	return filteredData;
 }
